pyproject.toml
[tool.poetry]
name = "gene-pathway-reasoning-agent"
version = "0.1.0"
description = "A production-ready, LLM-powered agent for gene-pathway reasoning with downstream analysis over a full database."
authors = ["Grigorios Maragkakis <grigorios.maragkakis@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
pydantic = "^2.11.2"
pydantic-settings = "^2.8.1"
python-dotenv = "^1.1.0"
openai = "^1.70.0"
lxml = "^5.3.1"
pandas = "^2.2.3"
networkx = "^3.4.2"
llama-index = "^0.12.28"
python-multipart = "^0.0.20"
markdown = "^3.7"
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"

[poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"